requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
Werkzeug==2.0.1
psd-tools==1.9.19
Pillow==8.3.1
python-dotenv==0.19.0
gunicorn==20.1.0
numpy==1.21.2
opencv-python==4.5.3.56
scikit-image==0.18.3
python-magic==0.4.24
pyinstaller==5.13.0 